Flip current working directory slashes
Since the cmd command takes in a quoted path, backslash directory
separators are incorrectly interpreted as escapes. To fix this, flip all
directory separators to forward slashes. In Win10, forward slashes are
recognized by the CD command, so everything works as expected, and the
new admin shell will execute in the launching directory as expected.