touch: removed
Had a puzzling error in a script that invoked the `touch` command.
Apparently this script ended up on my path before the actual .exe (from
the GOW toolset).

There's a nasty problem with shell scripts in that invoking a shell
script by default _chains_ to the script. That is, an invoked CMD script
does not return by default. To get the returning behavior after
invocation, you have to `call` the script. In other words,

    touch foo

invokes touch.cmd, which does not return to the calling script, but

    call touch foo

_does_ return.

You can partially address this by _always_ using `call` to invoke
commands (this works for .exe files as well), but that's a pain.

Since `touch` is such a common command, it seems that having `touch.cmd`
on your path is kind of a booby trap, and likely not worth the risk.
Thus, I've decided to drop this script from the collection.